+++ NEWS.md
@@ -1,4 +1,5 @@
 # Changes in version 1.5-0.1 (xx/xx/2016)
+* FIX: Error message for sorting with an unknown interest measure.
 
 
 # Changes in version 1.5-0 (09/23/2016)